# MQE: encode query plan node type as an integer, rather than a string