--- a/pkg/streamingpromql/planning/plan_test.go
+++ b/pkg/streamingpromql/planning/plan_test.go
@@ -91,6 +91,10 @@ func (t *testNode) Details() proto.Message {
 	panic("not supported")
 }
 
+func (t *testNode) NodeType() NodeType {
+	panic("not supported")
+}
+
 func (t *testNode) Children() []Node {
 	return t.children
 }