#### What this PR does

This PR improves the way query plans are encoded, replacing the use of a string like `type.googleapis.com/BinaryExpressionDetails` to identify the type of each node with an enum.

This significantly reduces the size of encoded payloads, which was the primary goal of this change. Previously, very simple expressions could result in very large encoded query plans (eg. `(a - b) + (c - d) + (e * f) + (g * h)`, which is 37 characters, would have resulted in a 1.12 KB encoded query plan).

It also significantly improves encoding and decoding time, although this was already fast in absolute terms.

Benchmark results:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql
cpu: Apple M1 Pro
                                                                                                                               │ original.txt  │            with-int.txt            │
                                                                                                                               │    sec/op     │   sec/op     vs base               │
PlanEncodingAndDecoding/foo/encode-10                                                                                             385.6n ± 18%   301.0n ± 1%  -21.95% (p=0.002 n=6)
PlanEncodingAndDecoding/foo/decode-10                                                                                             434.9n ±  1%   339.1n ± 1%  -22.03% (p=0.002 n=6)
PlanEncodingAndDecoding/foo{env="test",_region="us"}/encode-10                                                                    429.9n ±  0%   334.2n ± 1%  -22.25% (p=0.002 n=6)
PlanEncodingAndDecoding/foo{env="test",_region="us"}/decode-10                                                                    624.8n ±  0%   526.0n ± 0%  -15.81% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(foo)/encode-10                                                                                        673.2n ±  2%   480.6n ± 0%  -28.61% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(foo)/decode-10                                                                                        779.5n ±  1%   575.7n ± 0%  -26.14% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(env)_(foo)/encode-10                                                                              682.2n ±  1%   503.2n ± 1%  -26.23% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(env)_(foo)/decode-10                                                                              829.2n ±  0%   617.6n ± 0%  -25.52% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(foo)_+_sum(foo)/encode-10                                                                            1216.0n ±  0%   859.1n ± 1%  -29.35% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(foo)_+_sum(foo)/decode-10                                                                            1414.5n ±  0%   986.5n ± 1%  -30.26% (p=0.002 n=6)
PlanEncodingAndDecoding/foo_+_foo_+_foo_+_bar_+_foo/encode-10                                                                     2.123µ ±  0%   1.522µ ± 0%  -28.32% (p=0.002 n=6)
PlanEncodingAndDecoding/foo_+_foo_+_foo_+_bar_+_foo/decode-10                                                                     2.633µ ±  0%   1.872µ ± 0%  -28.90% (p=0.002 n=6)
PlanEncodingAndDecoding/a_+_b_+_c_+_d_+_e/encode-10                                                                               2.922µ ±  0%   2.150µ ± 2%  -26.44% (p=0.002 n=6)
PlanEncodingAndDecoding/a_+_b_+_c_+_d_+_e/decode-10                                                                               3.426µ ±  1%   2.534µ ± 3%  -26.01% (p=0.002 n=6)
PlanEncodingAndDecoding/(a_-_a)_+_(a_-_a)_+_(a_*_b)_+_(a_*_b)/encode-10                                                           3.399µ ±  9%   2.422µ ± 1%  -28.73% (p=0.002 n=6)
PlanEncodingAndDecoding/(a_-_a)_+_(a_-_a)_+_(a_*_b)_+_(a_*_b)/decode-10                                                           3.556µ ±  0%   2.521µ ± 1%  -29.10% (p=0.002 n=6)
PlanEncodingAndDecoding/(a_-_b)_+_(c_-_d)_+_(e_*_f)_+_(g_*_h)/encode-10                                                           5.109µ ±  0%   3.859µ ± 1%  -24.46% (p=0.002 n=6)
PlanEncodingAndDecoding/(a_-_b)_+_(c_-_d)_+_(e_*_f)_+_(g_*_h)/decode-10                                                           5.557µ ±  1%   4.143µ ± 0%  -25.44% (p=0.002 n=6)
PlanEncodingAndDecoding/123/encode-10                                                                                             345.7n ±  1%   266.3n ± 1%  -22.94% (p=0.002 n=6)
PlanEncodingAndDecoding/123/decode-10                                                                                             317.0n ±  1%   224.6n ± 0%  -29.16% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(rate(foo[5m]))/encode-10                                                                              950.1n ±  0%   684.7n ± 0%  -27.94% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(rate(foo[5m]))/decode-10                                                                             1183.0n ±  0%   864.4n ± 0%  -26.93% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(zone)_(label_replace(rate(foo[5m]),_"zone",_"$1",_"pod",_"ingester-zone-(.*)-\\d+"))/encode-10    2.188µ ±  0%   1.487µ ± 0%  -32.06% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(zone)_(label_replace(rate(foo[5m]),_"zone",_"$1",_"pod",_"ingester-zone-(.*)-\\d+"))/decode-10    2.724µ ±  0%   1.893µ ± 0%  -30.54% (p=0.002 n=6)
PlanEncodingAndDecoding/some_metric_*_on_(foo)_some_other_metric/encode-10                                                        970.3n ±  1%   715.0n ± 0%  -26.31% (p=0.002 n=6)
PlanEncodingAndDecoding/some_metric_*_on_(foo)_some_other_metric/decode-10                                                       1274.5n ±  2%   983.0n ± 1%  -22.88% (p=0.002 n=6)
PlanEncodingAndDecoding/some_metric_*_ignoring_(foo)_group_left_(bar)_some_other_metric/encode-10                                 982.4n ±  1%   719.6n ± 0%  -26.76% (p=0.002 n=6)
PlanEncodingAndDecoding/some_metric_*_ignoring_(foo)_group_left_(bar)_some_other_metric/decode-10                                 1.320µ ±  0%   1.020µ ± 0%  -22.73% (p=0.002 n=6)
PlanEncodingAndDecoding/(some_metric)[1m:1s]_offset_3s/encode-10                                                                  713.7n ±  1%   538.4n ± 0%  -24.56% (p=0.002 n=6)
PlanEncodingAndDecoding/(some_metric)[1m:1s]_offset_3s/decode-10                                                                  886.2n ±  1%   688.0n ± 2%  -22.37% (p=0.002 n=6)
PlanEncodingAndDecoding/max_over_time(rate(foo[5m:])[10m:])_+_max_over_time(rate(bar[5m:])[10m:])/encode-10                       3.738µ ±  1%   2.769µ ± 0%  -25.93% (p=0.002 n=6)
PlanEncodingAndDecoding/max_over_time(rate(foo[5m:])[10m:])_+_max_over_time(rate(bar[5m:])[10m:])/decode-10                       4.341µ ±  2%   3.272µ ± 2%  -24.63% (p=0.002 n=6)
PlanEncodingAndDecoding/label_join(foo,_"abc",_"-")_+_label_join(bar,_"def",_",")/encode-10                                       2.801µ ±  0%   1.990µ ± 2%  -28.97% (p=0.002 n=6)
PlanEncodingAndDecoding/label_join(foo,_"abc",_"-")_+_label_join(bar,_"def",_",")/decode-10                                       3.059µ ±  0%   2.162µ ± 0%  -29.34% (p=0.002 n=6)
geomean                                                                                                                           1.371µ         1.011µ       -26.23%

                                                                                                                               │ original.txt │           with-int.txt            │
                                                                                                                               │      B       │     B       vs base               │
PlanEncodingAndDecoding/foo/encode-10                                                                                              88.00 ± 0%   40.00 ± 0%  -54.55% (p=0.002 n=6)
PlanEncodingAndDecoding/foo{env="test",_region="us"}/encode-10                                                                    140.00 ± 0%   92.00 ± 0%  -34.29% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(foo)/encode-10                                                                                        165.00 ± 0%   64.00 ± 0%  -61.21% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(env)_(foo)/encode-10                                                                              180.00 ± 0%   79.00 ± 0%  -56.11% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(foo)_+_sum(foo)/encode-10                                                                              320.0 ± 0%   100.0 ± 0%  -68.75% (p=0.002 n=6)
PlanEncodingAndDecoding/foo_+_foo_+_foo_+_bar_+_foo/encode-10                                                                      541.0 ± 0%   176.0 ± 0%  -67.47% (p=0.002 n=6)
PlanEncodingAndDecoding/a_+_b_+_c_+_d_+_e/encode-10                                                                                682.0 ± 0%   242.0 ± 0%  -64.52% (p=0.002 n=6)
PlanEncodingAndDecoding/(a_-_a)_+_(a_-_a)_+_(a_*_b)_+_(a_*_b)/encode-10                                                            773.0 ± 0%   220.0 ± 0%  -71.54% (p=0.002 n=6)
PlanEncodingAndDecoding/(a_-_b)_+_(c_-_d)_+_(e_*_f)_+_(g_*_h)/encode-10                                                           1147.0 ± 0%   413.0 ± 0%  -63.99% (p=0.002 n=6)
PlanEncodingAndDecoding/123/encode-10                                                                                              77.00 ± 0%   30.00 ± 0%  -61.04% (p=0.002 n=6)
PlanEncodingAndDecoding/sum(rate(foo[5m]))/encode-10                                                                               247.0 ± 0%   100.0 ± 0%  -59.51% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(zone)_(label_replace(rate(foo[5m]),_"zone",_"$1",_"pod",_"ingester-zone-(.*)-\\d+"))/encode-10     682.0 ± 0%   301.0 ± 0%  -55.87% (p=0.002 n=6)
PlanEncodingAndDecoding/some_metric_*_on_(foo)_some_other_metric/encode-10                                                         303.0 ± 0%   157.0 ± 0%  -48.18% (p=0.002 n=6)
PlanEncodingAndDecoding/some_metric_*_ignoring_(foo)_group_left_(bar)_some_other_metric/encode-10                                  331.0 ± 0%   185.0 ± 0%  -44.11% (p=0.002 n=6)
PlanEncodingAndDecoding/(some_metric)[1m:1s]_offset_3s/encode-10                                                                   194.0 ± 0%   104.0 ± 0%  -46.39% (p=0.002 n=6)
PlanEncodingAndDecoding/max_over_time(rate(foo[5m:])[10m:])_+_max_over_time(rate(bar[5m:])[10m:])/encode-10                        865.0 ± 0%   367.0 ± 0%  -57.57% (p=0.002 n=6)
PlanEncodingAndDecoding/label_join(foo,_"abc",_"-")_+_label_join(bar,_"def",_",")/encode-10                                        693.0 ± 0%   267.0 ± 0%  -61.47% (p=0.002 n=6)
geomean                                                                                                                            327.7        136.3       -58.42%
```

We could make a similar change for function names: these are currently encoded as a string, but we could also use an enum to significantly reduce payload sizes. See #11955 for a sketch of this.

This is not a breaking change as nothing currently uses encoded query plans.

#### Which issue(s) this PR fixes or relates to

#11004 

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [covered by #10067] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.